refactor: allow building without deprecated code
