+++ .github/workflows/cd.yml
@@ -19,7 +19,6 @@ permissions:
 
 env:
   GOTOOLCHAIN: local
-  GOFLAGS: "-tags=nobadger,nomysql,nopgx"
 
 jobs:
   release:
