+++ .goreleaser.yml
@@ -19,6 +19,7 @@ builds:
     ldflags:
       - -X 'github.com/caddyserver/caddy/v2.CustomVersion=Mercure.rocks {{ .Version }} Caddy'
     tags:
+      - deprecated_transport
       - nobadger
       - nomysql
       - nopgx
@@ -40,6 +41,9 @@ builds:
     dir: cmd/mercure
     ldflags:
       - -s -w -X github.com/dunglas/mercure/common.version={{ .Version }} -X github.com/dunglas/mercure/common.commit={{ .ShortCommit }} -X github.com/dunglas/mercure/common.buildDate={{ .Date }}
+    tags:
+        - deprecated_transport
+        - deprecated_server
     goos:
       - linux
       - darwin
