+++ hub_notdeprecated.go
@@ -1,0 +1,5 @@
+//go:build !deprecated_server
+
+package mercure
+
+type deprecatedHub struct{}
