+++ hub_notdeprecated_test.go
@@ -1,0 +1,8 @@
+//go:build !deprecated_server
+
+package mercure
+
+import "testing"
+
+func setDeprecatedOptions(_ testing.TB, _ *Hub) {
+}
