+++ metrics_deprecated.go
@@ -1,0 +1,27 @@
+//go:build deprecated_server
+
+package mercure
+
+import (
+	"net/http"
+
+	"github.com/dunglas/mercure/common"
+	"github.com/gorilla/mux"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+)
+
+// Register configures the Prometheus registry with all collected metrics.
+//
+// Deprecated: use the Caddy server module or the standalone library instead.
+func (m *PrometheusMetrics) Register(r *mux.Router) {
+	// Metrics about current version
+	m.registry.MustRegister(common.AppVersion.NewMetricsCollector())
+
+	// Go-specific metrics about the process (GC stats, goroutines, etc.).
+	m.registry.MustRegister(prometheus.NewGoCollector())
+	// Go-unrelated process metrics (memory usage, file descriptors, etc.).
+	m.registry.MustRegister(prometheus.NewProcessCollector(prometheus.ProcessCollectorOpts{}))
+
+	r.Handle(metricsPath, promhttp.HandlerFor(m.registry.(*prometheus.Registry), promhttp.HandlerOpts{})).Methods(http.MethodGet)
+}
